+++ proxy/controller-grpc/Cargo.toml
@@ -1,0 +1,29 @@
+[package]
+name = "conduit-proxy-controller-grpc"
+version = "0.2.0"
+publish = false
+
+[features]
+default = []
+arbitrary = ["quickcheck"]
+
+[dependencies]
+convert = { path = "../convert" }
+
+bytes = "0.4"
+futures = "0.1"
+h2 = "0.1"
+http = "0.1"
+prost = "0.3.0"
+prost-derive = "0.3.0"
+prost-types = "0.3.0"
+
+tower-grpc = { git = "https://github.com/tower-rs/tower-grpc" }
+tower-h2   = { git = "https://github.com/tower-rs/tower-h2" }
+
+quickcheck = { version = "0.6", optional = true }
+
+
+[build-dependencies]
+tower-grpc-build = { git = "https://github.com/tower-rs/tower-grpc" }
+
